Resolve "Raise warning level and fix warnings on MSVC"

Closes #188

See merge request ped-dyn-emp/petrack!110

+++ src/codeMarkerItem.cpp
@@ -75,8 +75,8 @@ void CodeMarkerItem::paint(QPainter *painter, const QStyleOptionGraphicsItem */*
 
     if (mMainWindow->getCodeMarkerWidget()->showDetectedCandidates->isChecked())
     {
-        int nMarkers = mCorners.size();
-        int nRejected = mRejected.size();
+        int nMarkers = static_cast<int>(mCorners.size());
+        int nRejected = static_cast<int>(mRejected.size());
 
         cv::Point2f p0, p1;
         for(int i = 0; i < nMarkers; i++) // draws green square/ circle around head if person recognized
